use decimal for money

+++ csharp/ShoppingBasketDesignIdeas/ShoppingBasket1/BasketItem.cs
@@ -2,7 +2,7 @@ namespace ShoppingBasket1;
 
 public class BasketItem
 {
-    public BasketItem(string name, double price)
+    public BasketItem(string name, decimal price)
     {
         throw new NotImplementedException();
     }
